Refactor frontend to reduce duplication
