+++ web/src/components/App.vue
@@ -17,12 +17,6 @@
 
         <!-- Right aligned nav items -->
         <b-navbar-nav class="ml-auto">
-<!--
-          <b-nav-form>
-            <b-form-input size="sm" class="mr-sm-2" placeholder="Search"></b-form-input>
-            <b-button size="sm" class="my-2 my-sm-0" type="submit">Search</b-button>
-          </b-nav-form>
--->
         </b-navbar-nav>
       </b-collapse>
     </b-navbar>
